feat(home): move lecturer search onto home page and remove standalone search route
