--- src/routes/search.js
@@ -1,45 +1,0 @@
-'use strict'
-
-const express = require('express')
-const { connectToDatabase } = require('../models/db')
-const { searchLecturers } = require('../models/user_db')
-const { getSession } = require('../utils/session')
-
-const ROUTER = express.Router()
-
-/**
- * Renders the lecturer search page, filtering results by name, faculty, and school.
- * @param {import('express').Request} req - Express request object with optional query params `q`, `facultyId`, `schoolId`.
- * @param {import('express').Response} res - Express response object.
- */
-ROUTER.get('/', async (req, res) => {
-  const session = getSession(req)
-  if (!session) return res.redirect('/login')
-
-  const query = req.query.q?.trim() || ''
-  const facultyId = req.query.facultyId?.trim() || ''
-  const schoolId = req.query.schoolId?.trim() || ''
-
-  try {
-    await connectToDatabase()
-    const allLecturers = await searchLecturers({ universityId: session.universityId, query })
-
-    const faculties = [...new Set(allLecturers.map(lecturer => lecturer.facultyId).filter(Boolean))]
-    const schools = [...new Set(
-      allLecturers
-        .filter(lecturer => !facultyId || lecturer.facultyId === facultyId)
-        .map(lecturer => lecturer.schoolId)
-        .filter(Boolean)
-    )]
-    const lecturers = allLecturers.filter(lecturer =>
-      (!facultyId || lecturer.facultyId === facultyId) &&
-      (!schoolId || lecturer.schoolId === schoolId)
-    )
-
-    return res.render('search', { title: 'Find a Lecturer', session, lecturers, faculties, schools, query, facultyId, schoolId })
-  } catch {
-    return res.render('search', { title: 'Find a Lecturer', session, lecturers: [], faculties: [], schools: [], query, facultyId, schoolId })
-  }
-})
-
-module.exports = ROUTER
