+++ src/views/lecturer.ejs
@@ -26,7 +26,7 @@
         <% } %>
 
         <section>
-        <form action="/search" method="get">
+        <form action="/home" method="get">
             <button class="back_link" type="submit">Back to Search</button>
         </form>
         </section>
