--- src/views/search.ejs
@@ -1,69 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title><%= title %></title>
-    <link rel="stylesheet" href="/styles/main.css">
-</head>
-<body class="page page_home">
-    <main class="card card_wide">
-        <header>
-            <h1>Find a Lecturer</h1>
-        </header>
-
-        <section>
-            <form action="/search" method="get">
-                <label>
-                    Search
-                    <input type="text" name="q" value="<%= query%>" placeholder="Name or username">
-                </label>
-
-                <label>
-                    Faculty
-                    <select name="facultyId">
-                        <option value="">All faculties</option>
-                        <% faculties.forEach(function (faculty) {%>
-                            <option value="<%= faculty %>" <%= facultyId === faculty ? 'selected' : '' %>>
-                                <%= faculty %>
-                            </option>
-                        <% }) %>
-                    </select>
-                </label>
-
-                <label>
-                    School
-                    <select name="schoolId">
-                        <option value="">All schools</option>
-                        <% schools.forEach(function (school) { %>
-                            <option value="<%= school %>" <%= schoolId === school ? 'selected' : '' %>>
-                                <%= school %>
-                            </option>
-                        <% }) %>
-                    </select>
-                </label>
-
-                <button type="submit">Search</button>
-            </form>
-        </section>
-
-        <section class="results">
-            <% if (lecturers.length === 0) { %>
-                <p>No lecturers found.</p>
-            <% } else { %>
-                <% lecturers.forEach(function (lecturer) { %>
-                    <a href="/lecturer/<%= lecturer.username %>">
-                        <%= lecturer.firstName %> <%= lecturer.lastName %> &mdash; <%= lecturer.facultyId %>, <%= lecturer.schoolId %>
-                    </a>
-                <% }) %>
-            <% } %>
-        </section>
-
-        <section>
-            <form action="/home" method="get">
-                <button class="back_link" type="submit">Back to Home</button>
-            </form>
-        </section>
-    </main>
-</body>
-</html>
